fix: improve editor accessibility and refresh boundaries
